Consolidated interfaces into models
